CRP-6-2: Implement functions - Calculate booking amounts

+++ src/main/java/com/car_rental/project/model/Booking.java
@@ -49,7 +49,7 @@ public class Booking {
     private Long returnBranch;
 
     @Column(name = "booking_amount")
-    private int bookingAmount;
+    private Long bookingAmount;
 
 
     /*@OneToOne
